Moved window drawing to Console, to allow for easier window cleaning and stacking management
