+++ assets/text/help/rules.txt
@@ -1,0 +1,3 @@
+Escaping Eden is a sandbox game for all intents and purposes, however there are still rules that must be followed to ensure that the game stays fun and welcoming to new players.
+
+
